#1402 Remove ColumnSeries
