+++ Source/Examples/ExampleLibrary/Misc/XkcdExamples.cs
@@ -111,24 +111,24 @@ namespace ExampleLibrary
                 RenderingDecorator = rc => new XkcdRenderingDecorator(rc)
             };
 
-            var s1 = new ColumnSeries { Title = "Series 1", IsStacked = false, StrokeColor = OxyColors.Black, StrokeThickness = 1 };
-            s1.Items.Add(new ColumnItem { Value = 25 });
-            s1.Items.Add(new ColumnItem { Value = 137 });
-            s1.Items.Add(new ColumnItem { Value = 18 });
-            s1.Items.Add(new ColumnItem { Value = 40 });
+            var s1 = new BarSeries { Title = "Series 1", IsStacked = false, StrokeColor = OxyColors.Black, StrokeThickness = 1, XAxisKey = "x", YAxisKey = "y" };
+            s1.Items.Add(new BarItem { Value = 25 });
+            s1.Items.Add(new BarItem { Value = 137 });
+            s1.Items.Add(new BarItem { Value = 18 });
+            s1.Items.Add(new BarItem { Value = 40 });
 
-            var s2 = new ColumnSeries { Title = "Series 2", IsStacked = false, StrokeColor = OxyColors.Black, StrokeThickness = 1 };
-            s2.Items.Add(new ColumnItem { Value = 12 });
-            s2.Items.Add(new ColumnItem { Value = 14 });
-            s2.Items.Add(new ColumnItem { Value = 120 });
-            s2.Items.Add(new ColumnItem { Value = 26 });
+            var s2 = new BarSeries { Title = "Series 2", IsStacked = false, StrokeColor = OxyColors.Black, StrokeThickness = 1, XAxisKey = "x", YAxisKey = "y" };
+            s2.Items.Add(new BarItem { Value = 12 });
+            s2.Items.Add(new BarItem { Value = 14 });
+            s2.Items.Add(new BarItem { Value = 120 });
+            s2.Items.Add(new BarItem { Value = 26 });
 
-            var categoryAxis = new CategoryAxis { Position = AxisPosition.Bottom };
+            var categoryAxis = new CategoryAxis { Position = AxisPosition.Bottom, Key = "y" };
             categoryAxis.Labels.Add("Category A");
             categoryAxis.Labels.Add("Category B");
             categoryAxis.Labels.Add("Category C");
             categoryAxis.Labels.Add("Category D");
-            var valueAxis = new LinearAxis { Position = AxisPosition.Left, MinimumPadding = 0, MaximumPadding = 0.06, AbsoluteMinimum = 0 };
+            var valueAxis = new LinearAxis { Position = AxisPosition.Left, MinimumPadding = 0, MaximumPadding = 0.06, AbsoluteMinimum = 0, Key = "x" };
             model.Series.Add(s1);
             model.Series.Add(s2);
             model.Axes.Add(categoryAxis);
