--- Source/Examples/ExampleLibrary/Series/ColumnSeriesExamples.cs
@@ -1,15 +1,0 @@
-﻿// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="ColumnSeriesExamples.cs" company="OxyPlot">
-//   Copyright (c) 2014 OxyPlot contributors
-// </copyright>
-// --------------------------------------------------------------------------------------------------------------------
-
-namespace ExampleLibrary
-{
-    using OxyPlot.Series;
-
-    [Examples("ColumnSeries"), Tags("Series")]
-    public class ColumnSeriesExamples : BarAndColumnSeriesExamples<ColumnSeries, ColumnItem>
-    {
-    }
-}
