+++ Source/Examples/ExampleLibrary/Series/FinancialSeries/CandleStickSeriesExamples.cs
@@ -9,8 +9,6 @@ namespace ExampleLibrary
     using System;
     using System.Linq;
 
-    using ExampleLibrary.Utilities;
-
     using OxyPlot;
     using OxyPlot.Axes;
     using OxyPlot.Series;
@@ -147,12 +145,6 @@ namespace ExampleLibrary
             series.Items.Add(new HighLowItem(startTimeValue + 2, 99, 85, 93, 93));
             pm.Series.Add(series);
             return pm;
-        }
-
-        [Example("Simple CandleStickSeries example (transposed)")]
-        public static PlotModel SimpleExampleTransposed()
-        {
-            return SimpleExample().Transpose();
         }
 
         /// <summary>
