+++ Source/Examples/ExampleLibrary/Series/FinancialSeries/HighLowSeriesExamples.cs
@@ -40,22 +40,10 @@ namespace ExampleLibrary
             return model;
         }
 
-        [Example("HighLowSeries (transposed)")]
-        public static PlotModel HighLowSeriesTransposed()
-        {
-            return HighLowSeries().Transpose();
-        }
-
         [Example("HighLowSeries (reversed X Axis)")]
         public static PlotModel HighLowSeriesReversedXAxis()
         {
             return HighLowSeries().ReverseXAxis();
-        }
-
-        [Example("HighLowSeries (reversed X Axis, transposed)")]
-        public static PlotModel HighLowSeriesReversedXAxisTransposed()
-        {
-            return HighLowSeriesReversedXAxis().Transpose();
         }
 
         [Example("HighLowSeries (DateTime axis)")]
