+++ Source/Examples/ExampleLibrary/Series/FunctionSeriesExamples.cs
@@ -8,8 +8,6 @@ namespace ExampleLibrary
 {
     using System;
 
-    using ExampleLibrary.Utilities;
-
     using OxyPlot;
     using OxyPlot.Annotations;
     using OxyPlot.Axes;
@@ -22,12 +20,6 @@ namespace ExampleLibrary
         public static PlotModel SquareWave()
         {
             return CreateSquareWave(25);
-        }
-
-        [Example("Square wave (transposed)")]
-        public static PlotModel SquareWaveTransposed()
-        {
-            return SquareWave().Transpose();
         }
 
         private static PlotModel CreateSquareWave(int n = 25)
