+++ Source/Examples/ExampleLibrary/Series/StairStepSeriesExamples.cs
@@ -11,10 +11,7 @@ namespace ExampleLibrary
 {
     using System;
 
-    using ExampleLibrary.Utilities;
-
     using OxyPlot;
-    using OxyPlot.Axes;
     using OxyPlot.Series;
 
     /// <summary>
@@ -29,22 +26,10 @@ namespace ExampleLibrary
             return CreateExampleModel(new StairStepSeries());
         }
 
-        [Example("StairStepSeries (transposed)")]
-        public static PlotModel StairStepSeriesTransposed()
-        {
-            return StairStepSeries().Transpose();
-        }
-
         [Example("StairStepSeries with labels")]
         public static PlotModel StairStepSeriesWithLabels()
         {
             return CreateExampleModel(new StairStepSeries { LabelFormatString = "{1:0.00}" });
-        }
-
-        [Example("StairStepSeries with labels (transposed)")]
-        public static PlotModel StairStepSeriesWithLabelsTransposed()
-        {
-            return StairStepSeriesWithLabels().Transpose();
         }
 
         [Example("StairStepSeries with markers")]
