+++ Source/OxyPlot/Series/BarSeries/LinearBarSeries.cs
@@ -206,38 +206,49 @@ namespace OxyPlot.Series
         }
 
         /// <summary>
-        /// Applies an offset to a screen point.
-        /// </summary>
-        /// <param name="screenPoint">The screen point.</param>
-        /// <param name="offset">The offset.</param>
-        /// <returns>The translated screen point.</returns>
-        private static ScreenPoint Translate(ScreenPoint screenPoint, double offset)
-        {
-            return new ScreenPoint(screenPoint.X + offset, screenPoint.Y);
-        }
-
-        /// <summary>
         /// Find the index of a rectangle that contains the specified point.
         /// </summary>
         /// <param name="point">the target point</param>
         /// <returns>the rectangle index</returns>
         private int FindRectangleIndex(ScreenPoint point)
         {
-            var comparer = ComparerHelper.CreateComparer<OxyRect>(
-                (x, y) =>
-                {
-                    if (x.Right < point.X)
-                    {
-                        return -1;
-                    }
+            IComparer<OxyRect> comparer;
+            if (this.IsTransposed())
+            {
+                comparer = ComparerHelper.CreateComparer<OxyRect>(
+                    (x, y) =>
+                        {
+                            if (x.Bottom < point.Y)
+                            {
+                                return 1;
+                            }
 
-                    if (x.Left > point.X)
-                    {
-                        return 1;
-                    }
+                            if (x.Top > point.Y)
+                            {
+                                return -1;
+                            }
 
-                    return 0;
-                });
+                            return 0;
+                        });
+            }
+            else
+            {
+                comparer = ComparerHelper.CreateComparer<OxyRect>(
+                    (x, y) =>
+                        {
+                            if (x.Right < point.X)
+                            {
+                                return -1;
+                            }
+
+                            if (x.Left > point.X)
+                            {
+                                return 1;
+                            }
+
+                            return 0;
+                        });
+            }
 
             return this.rectangles.BinarySearch(0, this.rectangles.Count, new OxyRect(), comparer);
         }
@@ -251,6 +262,7 @@ namespace OxyPlot.Series
         private void RenderBars(IRenderContext rc, OxyRect clippingRect, List<DataPoint> actualPoints)
         {
             var widthOffset = this.GetBarWidth(actualPoints) / 2;
+            var widthVector = this.Orientate(new ScreenVector(widthOffset, 0));
 
             for (var pointIndex = 0; pointIndex < actualPoints.Count; pointIndex++)
             {
@@ -260,8 +272,8 @@ namespace OxyPlot.Series
                     continue;
                 }
 
-                var screenPoint = Translate(this.Transform(actualPoint), -widthOffset);
-                var basePoint = Translate(this.Transform(new DataPoint(actualPoint.X, 0)), widthOffset);
+                var screenPoint = this.Transform(actualPoint) - widthVector;
+                var basePoint = this.Transform(new DataPoint(actualPoint.X, 0)) + widthVector;
                 var rectangle = new OxyRect(basePoint, screenPoint);
                 this.rectangles.Add(rectangle);
                 this.rectanglesPointIndexes.Add(pointIndex);
